moved gol into one rust file
