+++ examples/src/game-of-life/src/templates/cell.pax
