+++ examples/src/game-of-life/src/templates/color_control.pax
