+++ examples/src/game-of-life/src/templates/lib.pax
