+++ examples/src/game-of-life/src/templates/speed_control.pax
